refactor: Entity 패키지 배치
