+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.Entity.Snapshot;
+import racingcar.Entity.Stadium;
+
 public class Application {
     public static void main(String[] args) {
         // 차 목록, 시도 횟수 입력 받기
